Create ModalCentered component
