+++ src/pages/Wallet/ConsolidateUTXOsModal.tsx
@@ -22,8 +22,7 @@ import styled, { useTheme } from 'styled-components'
 import Amount from '../../components/Amount'
 import Button from '../../components/Button'
 import InfoBox from '../../components/InfoBox'
-import { HeaderContent, HeaderLogo } from '../../components/Modal'
-import Modal from '../../components/Modal'
+import ModalCentered, { HeaderContent, HeaderLogo } from '../../components/ModalCentered'
 import { Section } from '../../components/PageComponents/PageContainers'
 import Spinner from '../../components/Spinner'
 
@@ -37,7 +36,7 @@ const ConsolidateUTXOsModal = ({ onConsolidateClick, onClose, fee }: Consolidate
   const theme = useTheme()
 
   return (
-    <Modal title="Consolidate UTXOs" onClose={onClose}>
+    <ModalCentered title="Consolidate UTXOs" onClose={onClose}>
       <HeaderContent>
         <HeaderLogo>
           <Codesandbox color={theme.global.accent} size={'70%'} strokeWidth={0.7} />
@@ -57,7 +56,7 @@ const ConsolidateUTXOsModal = ({ onConsolidateClick, onClose, fee }: Consolidate
           </Button>
         </Section>
       </HeaderContent>
-    </Modal>
+    </ModalCentered>
   )
 }
