+++ src/pages/Wallet/SendModalTransactionForm.tsx
@@ -25,7 +25,7 @@ import InfoBox from '../../components/InfoBox'
 import AddressSelect from '../../components/Inputs/AddressSelect'
 import AmountInput from '../../components/Inputs/AmountInput'
 import Input from '../../components/Inputs/Input'
-import { ModalFooterButton, ModalFooterButtons } from '../../components/Modal'
+import { ModalFooterButton, ModalFooterButtons } from '../../components/ModalCentered'
 import { useAddressesContext } from '../../contexts/addresses'
 import { checkAddressValidity } from '../../utils/addresses'
 import { MINIMAL_GAS_AMOUNT, MINIMAL_GAS_PRICE } from '../../utils/constants'
